requirements.txt
streamlit
pdfplumber
python-docx
pydantic
python-dotenv
google-genai
docx2pdf
